Create helpers to generate messages
