+++ disc/cog.py
@@ -1,21 +1,14 @@
 import logging
 from http.client import InvalidURL
 
-import discord
 from discord.ext import commands
 from discord_slash import SlashContext, cog_ext
-from discord_slash.model import ButtonStyle
-from discord_slash.utils.manage_components import create_actionrow, create_button
 
+from disc.message_generators import Descriptions, help_message, server_error_message
+from disc.message_generators.job_posting import invalid_url_message, job_info_message
 from job_api_wrapers import get_linkedin_job_info
-from utils import Colors
 
 _LOGGER = logging.getLogger(__name__)
-
-
-class Descriptions:
-    POST_JOB = "Parse a job post's information beautifully in the channel"
-    HELP = "Display the list of commands and their usages"
 
 
 class Slash(commands.Cog):
@@ -29,59 +22,18 @@ class Slash(commands.Cog):
         try:
             data = get_linkedin_job_info(job)
         except InvalidURL:
-            await ctx.send(
-                "Hi there! Seems like your URL or job ID is invalid. "
-                "Make sure that the URL is linking to a job post. "
-                "And keep in mind that currently Janus only supports LinkedIn job posts.",
-                hidden=True,
-            )
+            await ctx.send(**invalid_url_message())
             _LOGGER.info("Sent private message to user notifying given url is invalid.")
             return
 
         if data is None:
-            await ctx.send(
-                "Seems like we have an internal server error."
-                "Please try again later :(",
-                hidden=True,
-            )
+            await ctx.send(**server_error_message())
 
-        colors = Colors()
-        random_color = colors.random_color()
-        final_color = discord.Colour(random_color)
-
-        description = (
-            f"**{data['company']}**\n*{data['title']}*\n\n"
-            f"Location: {data['location']}.\n\n"
-            f"This job was posted __{data['posted_time_ago']}__.\n\n"
-        )
-
-        embed_content = discord.Embed(
-            title="Check out this job on LinkedIn!",
-            type="rich",
-            description=description,
-            color=final_color,
-        )
-        embed_content.set_thumbnail(url=f"{data['company_pic_url']}")
-
-        apply_button = create_button(
-            style=ButtonStyle.URL, label="Apply Now", url=data["url"]
-        )
-        action_row = create_actionrow(apply_button)
-
-        await ctx.send(embed=embed_content, components=[action_row])
+        await ctx.send(**job_info_message(data))
 
     @cog_ext.cog_slash(name="help", description=Descriptions.HELP)
     async def _help(self, ctx: SlashContext):
-        help_msg = f"""
-        `post-job` - {Descriptions.POST_JOB}
-        `help` - {Descriptions.HELP}
-        """
-        embed_content = discord.Embed(
-            title="Here's what you can do with this bot!",
-            type="rich",
-            description=help_msg,
-        )
-        await ctx.send(embed=embed_content)
+        await ctx.send(**help_message())
 
 
 def setup(bot):
